Extracted logic from fault tree holder main method - (Task #296)
---
Task #296: Refactor data structures in FaultTreeHolder
